% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distances.R
\name{AddDistances}
\alias{AddDistances}
\title{Return a data frame of meta data associatd with single cell data sets that
contains a distance from a reference}
\usage{
AddDistances(
  cell.data,
  signature.ref,
  path.glist,
  distance = 0,
  scramble = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{cell.data}{A Seurat object containing cells that are to be assigned a distance.}

\item{signature.ref}{Matrix of doubles. Reference data used to calculate the Rscore.}

\item{path.glist}{Directory. Directory of the gene list used for gene signature, stored in csv format.}

\item{distance}{Choice of model. 0-Mahalanobis 1-taxicab 2-z-score.}

\item{scramble}{Bool. To asses the effect of the gene signature on the Rscore, select a random set of genes and compute their distance from the reference data.}

\item{verbose}{Print error log.}
}
\description{
This function will compute the distance for each single cell. For
each cell in the reference and then return it in a data frame along with all the additional
single cell metadata
}
