% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotation.R
\name{AnnotateClonotypes}
\alias{AnnotateClonotypes}
\title{List clonotypes and the proportion of cells expressing a choice of marker genes
at normalized counts greater than the control}
\usage{
AnnotateClonotypes(
  Clonal_Obs,
  goi,
  goi.threshold = 0.975,
  index.control.input = 1,
  conditions = 1,
  path = "",
  names.spreadsheet = c(-1),
  preset = 1,
  threshold.entry = FALSE
)
}
\arguments{
\item{Clonal_Obs}{A Seurat object pre-processed with aocseq::CombineData.}

\item{goi}{A marker of interest.}

\item{goi.threshold}{Threshold of the goi normalized control counts per cell.}

\item{conditions}{Number of experiments within ClonalData object}

\item{path}{String. Directory of the table output.}

\item{names.spreadsheet}{List. Control and sample names.}

\item{preset}{Bool. If set to 1, the control is part of the dataset. If 0, a threshold
value for each condition must be set.}

\item{threshold.entry}{List. List of thresholds for each condition if preset is 0.}

\item{index.control}{List. Index of the control sequencing sample for each condition.}
}
\value{
A data frame containing clonotypes and summary statistics of transcript expression.
}
\description{
This function will read in Seurat objects processed by aocseq::CombineData
and generate a ranked spreadsheet of clonotypes grouped by CDR3 beta sequence.
The table displays the percentage of CD4 and CD8 cells expressing a choice of marker
genes at normalized counts greater than the control.
}
\concept{Annotation}
