% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distances.R
\name{ClassifyCells}
\alias{ClassifyCells}
\title{Return the Rscore of single cells in a query dataset compared to reference}
\usage{
ClassifyCells(
  output.array,
  normalized.array,
  signature.ref,
  Glist,
  distance = 0,
  scramble = FALSE,
  withSCT = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{output.array}{A Seurat object containing cells that are to be assigned an Rscore.}

\item{normalized.array}{A matrix of query data normalized counts.}

\item{signature.ref}{Matrix of doubles. Reference data used to calculate the Rscore.}

\item{Glist}{Vector. List of genes in the gene signature.}

\item{scramble}{Bool. To asses the effect of the gene signature on the Rscore, select a random set of genes and compute their distance from the reference data.}

\item{verbose}{Print progress bars and output}
}
\value{
A Seurat object containing metadata with the Rscore.
}
\description{
This function will compute the response score (Rscore) for each single cell in samples that have been activated
and sequenced. For each cell in the reference, we construct the positive definite covariance matrix
of the reference data set signature.ref. Takes as input two Seurat objects, one containing
the query cells and another containing the reference. A gene list
used to construct the distance and output path for plotting. Other parameters are listed for debugging,
but can be left
as default values.
there are different functions for different genesets
}
\concept{annotation}
