% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification.R
\name{ClassifyClonotypes}
\alias{ClassifyClonotypes}
\title{Combine single cell data and annotate with cell labels based on functionality}
\usage{
ClassifyClonotypes(
  Clonal_Obs,
  Clonotypes,
  within.sample = FALSE,
  method = "ZINB",
  goi = c("IFNG"),
  percentile = 0.01,
  path = "",
  path.glist = "",
  reference = matrix(nrow = 1, ncol = 1),
  verbose = TRUE
)
}
\arguments{
\item{Clonal_Obs}{A seurat object. Pre-processed with aocseq.}

\item{Clonotypes}{Table. An annotation table from the aocseq package.}

\item{method}{Vector of "ZINB", "Mahalanobis", "z-score" and "Taxi cab". "ZINB" can be used without a reference dataset.}

\item{goi}{Gene. Gene of interest (goi). For use with "ZINB" classification.}

\item{percentile}{Double. Percentile cutoff for distance based scoreing metrics.}

\item{path}{Directory for output tables.}

\item{verbose}{Print progress bars and output.}
}
\value{
A Seurat object list containing metadata and VDJ annotations.
}
\description{
This function will read in Seurat objects that have been processed with aocseq,
and accompanying annotation tables to classify clonotypes using distance scores.
Takes as input gene expression from scRNAseq and VDJ enrichment. Outputs a classification table
that lists samples and each clonotype classified.
}
\concept{annotation}
