% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import.R
\name{CombineData}
\alias{CombineData}
\title{Combine single cell data and annotate with cell labels based on functionality}
\usage{
CombineData(
  gex.path,
  marker.gene,
  vdj.path = c(),
  threshold.cutoff = 0.975,
  file.saved = "samples.rds",
  index.control = 1,
  n.samples = -1,
  c.index = -1,
  sample.name = -1,
  preset = 1,
  threshold.entry = 0,
  demultiplex = FALSE,
  n.hashtag.samples = 1,
  n.samples.ht = 1,
  tenX_conversion = "true",
  hashtags = 1,
  nFeature_RNA_lower = 100,
  nFeature_RNA_upper = 10000,
  nvariable_features = 3000,
  percent.mt_upper = 5,
  verbose = TRUE
)
}
\arguments{
\item{gex.path}{path to gene expression data in the format of cellranger feature barcode matrices}

\item{marker.gene}{list of marker genes used for specificity analysis}

\item{vdj.path}{path to VDJ expression data in the format of cellranger csv files}

\item{threshold.cutoff}{list of marker genes used for specificity analysis}

\item{file.saved}{directory for storing Seurat object RDS files}

\item{index.control}{index of the control sequencing sample}

\item{preset}{If set to 1, preset determines the threshold for cutoff if a control is included in the assay}

\item{threshold.entry}{Double. Sets the percentile above which gene expression is labelled as high.}

\item{demultiplex}{Boolean value used to indicate if the data was hashtagged}

\item{hashtags}{Integer represents the number of hashtags used (if any)}

\item{verbose}{Print progress bars and output}
}
\value{
A Seurat object list containing metadata and VDJ annotations.
}
\description{
This function will read in 10X genomics outputs from cellranger and combine
multiple Seurat objects, assigning labels based on additional tables stored
in csv format. Takes as input gene expression from scRNAseq and VDJ enrichment.
A gene list used to estimate specificity can be chosen and is application specific.
Other parameters are listed for debugging, but can be left as default values.
}
\concept{annotation}
