% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expression.R
\name{GetGeneSignature}
\alias{GetGeneSignature}
\title{Export differential gene expression from several different samples and phenotypes}
\usage{
GetGeneSignature(
  Clonal_Obs,
  clonotype.path,
  save.dir,
  goi,
  FClim = 10^(-3),
  specific.pval = 10^(-5),
  bystander.pval = 10^(-2),
  logfc.threshold = 0.01
)
}
\arguments{
\item{Clonal_Obs}{A Seurat object pre-processed with aocseq::CombineData.}

\item{clonotype.path}{Character array. Directory of an aocseq clonotype annotation table.}

\item{save.dir}{Directory for storing differentially expressed genes.}

\item{goi}{Character array. Gene name, a marker of interest.}

\item{verbose}{Print progress}
}
\value{
return an assay containing predicted expression value in the data
slot
}
\description{
This function will take a set of annotation spreadsheets and export differential
expression between different cell types labelled by their marker genes
Currently the phenotypic separation is CD4 and CD8, but the reader is encouraged
to generate as many subsets as needed for downstream analysis
}
\concept{integration}
