% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expression.R
\name{GetSpecificCells}
\alias{GetSpecificCells}
\title{Subsets a SEURAT object based on a list of meta data for the CD4 and CD8 subsets}
\usage{
GetSpecificCells(Clonal_Obs, expression, phenotype, tcrseq, goi_v)
}
\arguments{
\item{Clonal_Obs}{A Seurat object.}

\item{expression}{Choice of "high" or "unassigned".}

\item{phenotype}{CD4 or CD8 clasification.}

\item{tcrseq}{List of TCRs to mark clonotypes included in the analysis.}

\item{goi_v}{Gene of interest.}
}
\value{
A subset of a Seurat object processed with aocseq.
}
\description{
This function will read in Seurat objects processed by traceseq::AnnotateClonotypes
or any other software based on cell annotation and subset the SEURAT array accordingly
}
\concept{gene expression}
