% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import.R
\name{GMM_demux}
\alias{GMM_demux}
\title{Flexible implementation of the GMM demux algorithm}
\usage{
GMM_demux(s.name, data, hashtag_index, nameshashtags)
}
\arguments{
\item{s.name}{Name of the sample}

\item{data}{Input data. 10x object with hashtags}

\item{hashtag_index}{list containing positions of hashtags to be demultiplexed}

\item{nameshashtags}{Subsets of the sample for new orig.ident}
}
\value{
A Seurat object that is has
}
\description{
Demultiplex arbitrary number of hashtags
}
