% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import.R
\name{m_step}
\alias{m_step}
\title{Maximization Step of the EM Algorithm}
\usage{
m_step(x, posterior.df)
}
\arguments{
\item{x}{Input data.}

\item{posterior.df}{Posterior probability data.frame.}
}
\value{
Named list containing the mean (mu), variance (var), and mixing
  weights (alpha) for each component.
}
\description{
Update the Component Parameters
}
