% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/References.R
\name{MakeReference}
\alias{MakeReference}
\title{These are functions to produce reference datasets to compute a response score of
activated or perturbed T cells.}
\usage{
MakeReference(
  Perturbation,
  path.glist,
  TCR.list,
  sample.return = 3,
  n.inlist = 10,
  verbose = TRUE
)
}
\arguments{
\item{Perturbation}{List of Seurat objects. Seurat objects containing expression data processed.
by aocseq that will be used to score query T cell response.}

\item{path.glist}{path to the gene list of the gene signature.}

\item{TCR.list}{Character list. List of TCRs for clonotypes to include in the reference data.}

\item{sample.return}{Index of samples from the perturebation list that should be returned in the matrix.}

\item{n.inlist}{Number of genes from the gene signature that must have high expression for the cell to be included in the reference.}

\item{verbose}{Print progress bars and output.}
}
\value{
A reference matrix.
}
\description{
This function returns a matrix of normalized gene expression along the rows and cells
along the columns.
}
\concept{annotation.}
