% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{segmentPlot}
\alias{segmentPlot}
\title{Functions to plot data stored in clonal objects
at normalized counts greater than the control}
\usage{
segmentPlot(
  SummaryTable,
  names = c("S1", "S2"),
  segment.names = c("S1", "S2"),
  n.segments = 50,
  tune.gap = 0.005,
  segment.names.spacer = "16mm",
  percentage.spacer.1 = "14mm",
  percentage.spacer.2 = "13mm",
  segment.col = c("green", "blue"),
  ribbon.hue = "pink",
  path = "circos.pdf"
)
}
\arguments{
\item{SummaryTable}{A Seurat object pre-processed with aocseq::CombineData.}

\item{names}{A marker of interest.}

\item{n.segments}{Threshold of the goi normalized control counts per cell.}

\item{tune.gap}{List. Index of the control sequencing sample for each condition.}

\item{segment.names.spacer}{Number of experiments within ClonalData object}

\item{percentage.spacer.1}{String. Directory of the table output.}

\item{percentage.spacer.2}{List. Control and sample names.}

\item{segment.col}{Bool. If set to 1, the control is part of the dataset. If 0, a threshold
value for each condition must be set.}

\item{ribbon.hue}{List. List of thresholds for each condition if preset is 0.}
}
\value{
A data frame containing clonotypes and summary statistics of transcript expression.
}
\description{
This function will read in Seurat objects processed by aocseq::CombineData
and generate a ranked spreadsheet of clonotypes grouped by CDR3 beta sequence.
The table displays the percentage of CD4 and CD8 cells expressing a choice of marker
genes at normalized counts greater than the control.
}
\concept{Annotation}
